test(presets): merge registry.test.ts into registry.test.tsx
